requirements.txt
pycryptodome
argon2-cffi
cryptography
python-gnupg
